fix: correct type casting
