+++ src/Input.tsx
@@ -1,5 +1,6 @@
 import { removeKeysFromObject } from './helpers'
-import withKingpin, { WithKingpinType } from './withKingpin'
+import type { WithKingpinProps, WithKingpinType } from './types'
+import withKingpin from './withKingpin'
 import React, { ChangeEvent, FunctionComponent, InputHTMLAttributes } from 'react'
 
 type ReturnTypes = string | number | boolean
@@ -41,4 +42,4 @@ const KingpinInput = withKingpin<Props, ReturnTypes>((props: Props): JSX.Element
 
 KingpinInput.displayName = 'KingpinInput'
 
-export default KingpinInput as unknown as FunctionComponent<Props>
+export default KingpinInput as unknown as FunctionComponent<WithKingpinProps<ReturnTypes> & Props>
