+++ src/Select.tsx
@@ -1,5 +1,6 @@
 import { removeKeysFromObject } from './helpers'
-import withKingpin, { WithKingpinType } from './withKingpin'
+import type { WithKingpinProps, WithKingpinType } from './types'
+import withKingpin from './withKingpin'
 import React, { FunctionComponent, SelectHTMLAttributes } from 'react'
 
 type Props = SelectHTMLAttributes<HTMLSelectElement> & WithKingpinType<string>
@@ -20,4 +21,4 @@ const KingpinSelect = withKingpin<Props, string>(
 )
 KingpinSelect.displayName = 'KingpinSelect'
 
-export default KingpinSelect as unknown as FunctionComponent<Props>
+export default KingpinSelect as unknown as FunctionComponent<WithKingpinProps<string> & Props>
